rename component folder to components
