+++ src/App.js
@@ -1,6 +1,6 @@
-import Main from './component/Main'
-import Footer from './component/Footer'
-import Header from './component/Header'
+import Main from './components/Main'
+import Footer from './components/Footer'
+import Header from './components/Header'
 import './styles/styles.css'
 
 function App() {
